+++ R/fetch_functions.R
@@ -1,21 +1,21 @@
 batchData <- function(ga, ga_pars){
   ## TODO: max results currently ignored if over 10000, change that.
   ga_pars$`max-results` <- 10000
-  message("Finding number of results to fetch")
+  myMessage("Finding number of results to fetch", level = 2)
   check_data <- ga(pars_arguments = ga_pars)
   batches <- attr(check_data, "totalResults") %/% 10000
-
+  
   anyBatchSampled <- attr(check_data, 'containsSampledData')
   samplePercent   <- attr(check_data, "samplePercent")
-
+  
   if(batches > 0){
-    message("Batching data into [", batches+1, "] calls.")
-
+    myMessage("Batching data into [", batches+1, "] calls.", level = 3)
+    
     batch_start_index <- seq(1, (batches+1)*10000, 10000)
-
+    
     anyBatchSampled <- FALSE
     samplePercent   <- 0
-
+    
     bf <- function(batch_data){
       lapply(batch_data, function(the_data) {
         if(attr(the_data, 'containsSampledData')) anyBatchSampled <<- TRUE
@@ -23,25 +23,25 @@ batchData <- function(ga, ga_pars){
       })
       batch_data
     }
-
+    
     the_data <- googleAuthR::gar_batch_walk(ga,
                                             batch_start_index,
                                             gar_pars = ga_pars,
                                             pars_walk = "start-index",
                                             batch_function = bf)
-
+    
   } else{
     the_data <- check_data
   }
-
-  message("All data found. [",
+  
+  myMessage("All data found. [",
           NROW(the_data), "] total results out of a possible [",
-          attr(the_data, "totalResults") , "]" )
-
+          attr(the_data, "totalResults") , "]", level = 3)
+  
   attr(the_data, "containsSampledData") <- anyBatchSampled
   attr(the_data, "samplePercent") <- samplePercent / (batches + 1)
   the_data
-
+  
 }
 
 walkData <- function(ga, ga_pars, start, end){
@@ -49,13 +49,13 @@ walkData <- function(ga, ga_pars, start, end){
     seq(as.Date(start, format="%Y-%m-%d"),
         as.Date(end, format="%Y-%m-%d"),
         by=1))
-
+  
   ga_pars$samplingLevel <- "HIGHER_PRECISION"
   ga_pars$`max-results` <- 10000
-
+  
   anyBatchSampled <- FALSE
   samplePercent   <- 0
-
+  
   bf <- function(batch_data){
     lapply(batch_data, function(the_data) {
       if(attr(the_data, 'containsSampledData')) anyBatchSampled <<- TRUE
@@ -63,22 +63,22 @@ walkData <- function(ga, ga_pars, start, end){
     })
     batch_data
   }
-
+  
   walked_data <- googleAuthR::gar_batch_walk(ga,
                                              dates,
                                              gar_pars = ga_pars,
                                              pars_walk = c("start-date", "end-date"),
                                              batch_function = bf)
-
-  message("Walked through all dates. Total Results: [", NROW(walked_data), "]")
+  
+  myMessage("Walked through all dates. Total Results: [", NROW(walked_data), "]", level = 3)
   attr(walked_data, "dateRange") <- list(startDate = start, endDate = end)
   attr(walked_data, "totalResults") <- NROW(walked_data)
   attr(walked_data, "samplingLevel") <- "HIGHER_PRECISION, WALKED"
   attr(walked_data, "containsSampledData") <- anyBatchSampled
   attr(walked_data, "samplePercent") <- samplePercent / length(dates)
-
+  
   walked_data
-
+  
 }
 
 
@@ -89,7 +89,7 @@ loop_ids <- function(id, ga_pars, samplingLevel, max_results, ga){
     
     ## if walk through results then split up dates and walk through date ranges
     if(samplingLevel %in% "WALK"){
-      message("Walking through data.")
+      myMessage("Walking through data.", level = 3)
       the_data <- walkData(ga, ga_pars, ga_pars$`start-date`, ga_pars$end)
       
     } else {
@@ -105,7 +105,7 @@ loop_ids <- function(id, ga_pars, samplingLevel, max_results, ga){
     
     if(length(id) > 1){
       ## for multiple id's, a list of dataframes.
-      message("Multiple IDs")
+      myMessage("Multiple IDs", level = 2)
       id_name <- attr(the_data, "profileInfo")$profileId
       
       all_data[[id_name]] <- the_data